+++ kubejs/startup_scripts/global.js
@@ -46,11 +46,12 @@ global.removedItems = [
 	'thermal:machine_crafter',
 	'thermal:machine_press',
 	'thermal:potato_block',
+	'thermal:satchel',
 	'thermal:sushi_maki',
 	/^thermal:dynamo_.*/,
 	/^thermal:flax.*/,
 	/^thermal:onion.*/,
-	/^thermal:press_.*_die$/,
+	/^thermal:press_(?!gear).*_die$/,
 	/^thermal:rice.*/,
 	/^thermal:tomato.*/
 ];